requirements.txt
Django
psycopg2
requests
social-auth-app-django
django-allauth
pandas==1.4.2
openpyxl
gunicorn
django-dbconn-retry
scikit-learn==1.1.3
numpy==1.23.5
neo4j